refactor(clean/run): estrai SQL execution in clean/sql_execute.py
Issue #140 step 1: separa il boundary SQL/DuckDB execution dall'orchestrazione.

- toolkit/clean/sql_execute.py (nuovo): _run_sql, _normalize_output_profile
- toolkit/clean/run.py: importa da sql_execute; rimossi _run_sql e _normalize_output_profile locali

sql_dry_run.py non richiede modifiche (importa ancora _load_clean_sql da clean.run).

371/371 test passano. ruff clean.
